% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idxAnno.R
\name{buildCustomIdx}
\alias{buildCustomIdx}
\title{Custom Gene Set Collection Index}
\usage{
buildCustomIdx(entrezIDs, gsets, anno = NULL, label = "custom",
  name = "User-Defined Gene Sets", species = "Human", min.size = 1)
}
\arguments{
\item{entrezIDs}{character, a vector that stores the Entrez Gene IDs tagged 
in your dataset. 
The order of the Entrez Gene IDs should match those of the count/expression 
matrix row names.}

\item{gsets}{list, list of gene sets. Each gene set is character vector of 
Enterz IDs. 
The names of the list should match the GeneSet column in the \code{anno} 
argument (if it is provided).}

\item{anno}{list, dataframe that stores a detailed annotation for each  gene 
set. 
Some of its fields can be ID, GeneSet, PubMed, URLs, etc. The GeneSet field 
is mandatory and
should have the same names as the \code{gsets}' names.}

\item{label}{character,a unique id that identifies the collection of gene 
sets}

\item{name}{character,the collection name to be used in the EGSEA report}

\item{species}{character, determine the organism of selected gene sets: 
"human", "mouse" or "rat".}

\item{min.size}{integer, the minium number of genes required in a testing 
gene set}
}
\value{
indexed gene set annotation that can be used with other functions in 
the package.
Each annotation is a list of seven elements: \code{original} stores the 
original gene sets, 
\code{idx} stores the indexed gene sets,  \code{anno} that stores detailed 
annotation for each 
gene set, \code{label} a unique id that identifies the collection of gene 
sets, 
\code{featureIDs} stores the entrezIDs used in building the annotation, 
\code{species}
 stores that organism name of gene sets and \code{name}  stores the 
collection name 
to be used in the EGSEA report.
}
\description{
It creates gene set collections from a given list of gene sets 
to be used 
for the EGSEA analysis.
}
\details{
It indexes newly created gene sets and attach gene set annotation 
if provided.
}
\examples{
library(EGSEAdata) 
data(il13.data)
v = il13.data$voom
data(kegg.pathways)
gsets = kegg.pathways$human$kg.sets[1:50]
gs.annot = buildCustomIdx(entrezIDs=rownames(v$E), gsets= gsets, 
species="human")
class(gs.annot)


}

