% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idxAnno.R
\name{buildGeneSetDBIdx}
\alias{buildGeneSetDBIdx}
\title{Gene Set Collection Indexes from the GeneSetDB Database}
\usage{
buildGeneSetDBIdx(entrezIDs, species, geneSets = "all", min.size = 1)
}
\arguments{
\item{entrezIDs}{character, a vector that stores the Entrez Gene IDs tagged 
in your dataset. 
The order of the Entrez Gene IDs should match those of the count/expression 
matrix row names.}

\item{species}{character, determine the organism of selected gene sets: 
"human", "mouse" or "rat".}

\item{geneSets}{character, a vector determines which gene set collections
are loaded from the GeneSetDB.
It takes "all", "gsdbdis", "gsdbgo", "gsdbdrug", "gsdbpath" or "gsdbreg". 
"all" includes all the GeneSetDB collections.  
"gsdbdis" is to load the disease collection, "gsdbgo" to load the GO terms collection,
 "gsdbdrug" to load the drug/chemical collection, 
"gsdbpath" to load the pathways collection and "gsdbreg" to load the gene regulation
collection.}

\item{min.size}{integer, the minium number of genes required in a testing 
gene set}
}
\value{
indexed gene set annotation that can be used with other functions in 
the package.
Each annotation is a list of seven elements: \code{original} stores the 
original gene sets, 
\code{idx} stores the indexed gene sets,  \code{anno} that stores detailed 
annotation for each 
gene set, \code{label} a unique id that identifies the collection of gene 
sets, 
\code{featureIDs} stores the entrezIDs used in building the annotation, 
\code{species}
 stores that organism name of gene sets and \code{name}  stores the 
collection name 
to be used in the EGSEA report.
}
\description{
It prepares the GeneSetDB gene set collections to be used for 
the EGSEA analysis.
}
\details{
It indexes the GeneSetDB gene sets and loads gene set annotation.
}
\examples{
library(EGSEAdata)
data(il13.data)
v = il13.data$voom
gs.annots = buildGeneSetDBIdx(entrezIDs=rownames(v$E), species="human")
names(gs.annots)


}

