% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idxAnno.R
\name{buildIdx}
\alias{buildIdx}
\title{Generate Gene Set Collection Indexes from the MSigDB and KEGG 
Databases}
\usage{
buildIdx(entrezIDs, species = "human", msigdb.gsets = "all",
  gsdb.gsets = "none", kegg.updated = FALSE, kegg.exclude = c(),
  min.size = 1, rdata.dir = NULL)
}
\arguments{
\item{entrezIDs}{character, a vector that stores the Entrez Gene IDs tagged 
in your dataset. 
The order of the Entrez Gene IDs should match those of the count/expression 
matrix row names.}

\item{species}{character, determine the organism of selected gene sets: 
"human", "mouse" or "rat".}

\item{msigdb.gsets}{character, a vector determines which gene set 
collections should be used from MSigDB.
It can take values from this list: "h", "c1", "c2", "c3", "c4", "c5", 
"c6","c7". "h" and "c1"
are human specific. If "all", all available gene set collections are loaded. 
If "none", 
MSigDB collections are excluded.}

\item{gsdb.gsets}{character, a vector determines which gene set collections
are loaded from the GeneSetDB.
It takes "none", "all", "gsdbdis", "gsdbgo", "gsdbdrug", "gsdbpath" or "gsdbreg". 
"none" excludes the GeneSetDB collections.
"all" includes all the GeneSetDB collections.  
"gsdbdis" to load the disease collection, "gsdbgo" to load the GO terms collection,
 "gsdbdrug" to load the drug/chemical collection, 
"gsdbpath" to load the pathways collection and "gsdbreg" to load the gene regulation
collection.}

\item{kegg.updated}{logical, set to TRUE if you want to download the most 
recent KEGG pathways.}

\item{kegg.exclude}{character, vector used to exclude KEGG pathways of 
specific type(s): 
Disease, Metabolism, Signaling. If "all", none fo the KEGG collections is 
included.}

\item{min.size}{integer, the minium number of genes required in a testing 
gene set}

\item{rdata.dir}{character, directory from which the MSigDB collections are 
loaded. 
If NULL, EGSEA tries to load the data from \pkg{EGSEAdata}.}
}
\value{
indexed gene set annotation that can be used with other functions in 
the package.
Each annotation is a list of seven elements: \code{original} stores the 
original gene sets, 
\code{idx} stores the indexed gene sets,  \code{anno} that stores detailed 
annotation for each 
gene set, \code{label} a unique id that identifies the collection of gene 
sets, 
\code{featureIDs} stores the entrezIDs used in building the annotation, 
\code{species}
 stores that organism name of gene sets and \code{name}  stores the 
collection name 
to be used in the EGSEA report.
}
\description{
It prepares the MSigDB and KEGG gene set collections to be used 
for the EGSEA analysis.
}
\details{
It indexes the MSigDB and KEGG gene sets and loads gene set 
annotation.
}
\examples{
library(EGSEAdata)
data(il13.data)
v = il13.data$voom
gs.annots = buildIdx(entrezIDs=rownames(v$E), species="human",
         msigdb.gsets = c("h", "c2"),
         kegg.exclude = c("Metabolism"))
names(gs.annots)
}

