% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idxAnno.R
\name{buildKEGGIdx}
\alias{buildKEGGIdx}
\title{Gene Set Collection Index from the KEGG Database}
\usage{
buildKEGGIdx(entrezIDs, species = "human", min.size = 1, updated = FALSE,
  rdata.dir = NULL)
}
\arguments{
\item{entrezIDs}{character, a vector that stores the Entrez Gene IDs tagged 
in your dataset. 
The order of the Entrez Gene IDs should match those of the count/expression 
matrix row names.}

\item{species}{character, determine the organism of selected gene sets: 
"human", "mouse" or "rat".}

\item{min.size}{integer, the minium number of genes required in a testing 
gene set}

\item{updated}{logical, set to TRUE if you want to download the most recent 
KEGG pathways.}

\item{rdata.dir}{character, directory from which the KEGG pathway 
collections are loaded. 
If NULL, EGSEA tries to load the data from \pkg{EGSEAdata}.}
}
\value{
indexed gene set annotation that can be used with other functions in 
the package.
Each annotation is a list of seven elements: \code{original} stores the 
original gene sets, 
\code{idx} stores the indexed gene sets,  \code{anno} that stores detailed 
annotation for each 
gene set, \code{label} a unique id that identifies the collection of gene 
sets, 
\code{featureIDs} stores the entrezIDs used in building the annotation, 
\code{species}
 stores that organism name of gene sets and \code{name}  stores the 
collection name 
to be used in the EGSEA report.
}
\description{
It prepares the KEGG pathway collection to be used for the 
EGSEA analysis.
}
\details{
It indexes the KEGG pathway gene sets and loads gene set annotation.
}
\examples{
library(EGSEAdata)
data(il13.data)
v = il13.data$voom
gs.annots = buildKEGGIdx(entrezIDs=rownames(v$E), species="human")

}

