% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idxAnno.R
\name{buildMSigDBIdxEZID}
\alias{buildMSigDBIdxEZID}
\title{Gene Set Collection Indexes from the MSigDB Database}
\usage{
buildMSigDBIdxEZID(entrezIDs, geneSets = NULL, species = "Homo \\nsapiens",
  min.size = 1, rdata.dir = NULL)
}
\arguments{
\item{entrezIDs}{character, a vector that stores the Entrez Gene IDs tagged 
in your dataset. 
The order of the Entrez Gene IDs should match those of the count/expression 
matrix row names.}

\item{geneSets}{character, a vector determines which gene set collections 
should be used.
It can take values from this list: "h", "c1", "c2", "c3", "c4", "c5", 
"c6","c7". "h" and "c1"
are human specific. If NULL, all available gene set collections are loaded.}

\item{species}{character, determine the organism of selected gene sets: 
"human", "mouse" or "rat".}

\item{min.size}{integer, the minium number of genes required in a testing 
gene set}

\item{rdata.dir}{character, directory from which the MSigDB collections are 
loaded. 
If NULL, EGSEA tries to load the data from \pkg{EGSEAdata}.}
}
\value{
indexed gene set annotation that can be used with other functions in 
the package.
Each annotation is a list of seven elements: \code{original} stores the 
original gene sets, 
\code{idx} stores the indexed gene sets,  \code{anno} that stores detailed 
annotation for each 
gene set, \code{label} a unique id that identifies the collection of gene 
sets, 
\code{featureIDs} stores the entrezIDs used in building the annotation, 
\code{species}
 stores that organism name of gene sets and \code{name}  stores the 
collection name 
to be used in the EGSEA report.
}
\description{
It prepares the MSigDB gene set collections to be used for the 
EGSEA analysis.
}
\details{
It indexes the MSigDB gene sets and loads gene set annotation.
}
\examples{
library(EGSEAdata)
data(il13.data)
v = il13.data$voom
gs.annots = buildMSigDBIdxEZID(entrezIDs=rownames(v$E), geneSets=c("h", 
"c2"), species="human")
names(gs.annots)
}

