% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egsea.R
\docType{package}
\name{EGSEA-package}
\alias{EGSEA}
\alias{EGSEA-package}
\title{Ensemble of Gene Enrichment Analysis (EGSEA)}
\description{
This packages provides the implementatino of the EGSEA 
algorithm 
and addition functions to help perform GSE analysis
}
\author{
Monther Alhamdoosh, Milica Ng and Matthew Ritchie
}

