% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egsea.R
\name{egsea.base}
\alias{egsea.base}
\title{EGSEA Base GSE Methods}
\usage{
egsea.base()
}
\value{
It returns a character vector of supported GSE methods.
}
\description{
It lists the supported GSEA methods
}
\examples{
egsea.base()
}

