% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egsea.R
\name{egsea.cnt}
\alias{egsea.cnt}
\title{Ensemble of Gene Set Enrichment Analyses Function}
\usage{
egsea.cnt(counts, group, design = NULL, contrasts = NULL, logFC = NULL,
  gs.annots, symbolsMap = NULL, baseGSEAs = egsea.base(), minSize = 2,
  display.top = 20, combineMethod = "fisher", combineWeights = NULL,
  sort.by = "p.adj", egsea.dir = NULL, kegg.dir = NULL,
  logFC.cutoff = 0, fdr.cutoff = 0.05, sum.plot.axis = "p.adj",
  sum.plot.cutoff = NULL, vote.bin.width = 5, num.threads = 4,
  report = TRUE, keep.base = FALSE, verbose = FALSE, keep.limma = TRUE,
  keep.set.scores = FALSE)
}
\arguments{
\item{counts}{double, an K x M numeric matrix of read counts where genes are the rows 
and samples are
the columns.}

\item{group}{character, vector or factor giving the experimental 
group/condition for each sample/library}

\item{design}{double, an M x N numeric matrix giving the design matrix of the linear 
model fitting.}

\item{contrasts}{double, an N x L matrix indicates the contrasts of the 
linear model coefficients for
which the test is required. N is the number of columns of the 
design matrix and L is 
number of contrasts. Can be also a vector of integers that specify the columns of the
design matrix.}

\item{logFC}{double, an K x L matrix indicates the log2 fold change of each 
gene for each contrast. 
K is the number of genes included in the analysis. If logFC=NULL, the logFC 
values are 
estimated using the  \code{\link[limma]{eBayes}} for each contrast.}

\item{gs.annots}{list, list of objects of class GSCollectionIndex. It is generated 
using one of these functions: 
 \code{\link{buildIdx}}, \code{\link{buildMSigDBIdx}}, 
\code{\link{buildKEGGIdx}},
\code{\link{buildGeneSetDBIdx}}, and \code{\link{buildCustomIdx}}.}

\item{symbolsMap}{dataframe, an K x 2 matrix stores the gene symbol of each 
Entrez Gene ID. It 
is used for the heatmap visualization. The order of rows should match that 
of the 
\strong{counts}. Default symbolsMap=NULL.}

\item{baseGSEAs}{character, a vector of the gene set tests that should be 
included in the 
ensemble. Type  \code{\link{egsea.base}} to see the supported GSE methods. 
By default, all
supported methods are used.}

\item{minSize}{integer, the minimum size of a gene set to be included in the 
analysis. 
Default minSize= 2.}

\item{display.top}{integer, the number of top gene sets to be displayed in 
the EGSEA report.
You can always access the list of all tested gene sets using the returned 
gsa list. 
Default is 20.}

\item{combineMethod}{character, determines how to combine p-values from 
different 
GSEA method. Type  \code{\link{egsea.combine}}() to see supported methods.}

\item{combineWeights}{double, a vector determines how different GSEA methods 
will be weighted. 
Its values should range between 0 and 1. This option is not supported 
currently.}

\item{sort.by}{character, determines how to order the analysis results in 
the stats table. Type  
\code{\link{egsea.sort}}() to see all available options.}

\item{egsea.dir}{character, directory into which the analysis results are 
written out.}

\item{kegg.dir}{character, the directory of KEGG pathway data file (.xml) 
and image file (.png). 
Default kegg.dir=paste0(egsea.dir, "/kegg-dir/").}

\item{logFC.cutoff}{numeric, cut-off threshold of logFC and is used for 
the calculation of Sginificance Score 
and Regulation Direction. Default logFC.cutoff=0.}

\item{fdr.cutoff}{numeric, cut-off threshold of DE genes and is used
for the calculation of Significance Score and Regulation Direction. 
Default fdr.cutoff = 0.05.}

\item{sum.plot.axis}{character, the x-axis of the summary plot. All the 
values accepted by the 
\strong{sort.by} parameter can be used. Default sum.plot.axis="p.value".}

\item{sum.plot.cutoff}{numeric, cut-off threshold to filter the gene sets of 
the summary plots
based on the values of the \strong{sum.plot.axis}. Default 
sum.plot.cutoff=NULL.}

\item{vote.bin.width}{numeric, the bin width of the vote ranking. Default 
vote.bin.width=5.}

\item{num.threads}{numeric, number of CPU cores to be used. Default 
num.threads=2.}

\item{report}{logical, whether to generate the EGSEA interactive report. It 
takes longer time
to run. Default is True.}

\item{keep.base}{logical, whether to write out the results of the 
individual GSE methods.
 Default FALSE.}

\item{verbose}{logical, whether to print out progress messages and warnings.}

\item{keep.limma}{logical, whether to store the results of the limma analysis
in the EGSEAResults object.}

\item{keep.set.scores}{logical, whether to calculate the gene set enrichment scores
per sample for the methods that support this option, i.e., "ssgsea".}
}
\value{
A list of elements, each with two/three elements that store the top 
gene sets and the detailed analysis
results for each contrast and the comparative analysis results.
}
\description{
This is the main function to carry out gene set enrichment 
analysis using the
 EGSEA algorithm. This function is aimed to use the raw count matrix to 
perform the EGSEA analysis.
}
\details{
EGSEA, an acronym for \emph{Ensemble of Gene Set Enrichment 
Analyses}, utilizes the 
analysis results of eleven prominent GSE algorithms from the literature to 
calculate 
collective significance scores for gene sets. These methods include: 
\pkg{ora}, 
\pkg{globaltest}, \pkg{plage}, \pkg{safe}, \pkg{zscore}, \pkg{gage}, 
\pkg{ssgsea},
\pkg{roast}, \pkg{fry}, \pkg{padog}, \pkg{camera} and \pkg{gsva}. 
The ora, gage, camera and gsva methods depend on a competitive null 
hypothesis while the 
remaining seven methods are based on a self-contained hypothesis. 
Conveniently, the 
algorithm proposed here is not limited to these eleven GSE methods and new 
GSE tests 
can be easily integrated into the framework. This function takes the raw 
count matrix, 
the experimental group of each sample, the design matrix and the contrast 
matrix as parameters. 
It performs TMM normalization and then applies \link[limma]{voom} to 
calculate the logCPM and weighting factors. 
The results of EGSEA can be seen using the \code{\link{topSets}} function. \cr 
 \cr
 EGSEA report is an interactive HTML report that is generated if \code{report=TRUE} to
enable a swift navigation through the results of an EGSEA analysis. The following pages 
are generated for each gene set collection and contrast/comparison: \cr
1. Stats Table page shows the detailed statistics of the EGSEA analysis for the 
\code{display.top} gene sets. It shows the EGSEA scores, individual rankings and 
additional annotation for each gene set. Hyperlinks to the source of each gene set
can be seen in this table when they are available. The "Direction" column shows the regulation
direction of a gene set which is calculated based on the \code{logFC}, which is
either calculated from the limma differential expression analysis or provided by the user. 
The \code{logFC.cutoff} and \code{fdr.cutoff} are applied for this calculation.
 The calculations of the EGSEA 
scores can be seen in the references section. The method \code{topSets} can be used to
generate custom Stats Table. \cr
2. Heatmaps page shows the heatmaps of the gene fold changes for the gene sets that are
presented in the Stats Table page. Red indicates up-regulation
while blue indicates down-regulation. Only genes that appear in the input expression/count 
matrix are visualized in the heat map. Gene names are coloured based on their 
statistical significance in the \code{limma} differential expression analysis. 
The "Interpret Results" link below each heat map allows the user to download the 
original heat map values along with additional statistics from \code{limma} DE analysis (
if available) so that they can be used to perform further analysis in R, e.g., customizing 
the heat map visualization. Additional heat maps can be generated and customized
 using the method \code{plotHeatmap}. \cr
3. Summary Plots page  shows the methods ranking plot along with the summary plots of 
EGSEA analysis. The method plot uses multidimensional scaling (MDS) to visualize the 
ranking of individual methods on a given gene set collection. The summary plots are 
bubble plots that visualize the distribution of gene sets based on the EGSEA
Significance Score and another EGSEA score (default, p-value). 
Two summary plots are generated: ranking and directional plots. Each gene set is 
reprersented with a bubble which is coloured based on the EGSEA ranking (in ranking
plots ) or gene set regulation direction (in directional plots) and sized based on the 
gene set cardinality (in ranking plots) or EGSEA Significance score (in directional plots).
Since the EGSEA "Significance Score" is proportional to the p-value and the 
absolute fold changes, it could be useful to highlight gene sets that
have high Significance scores. The blue labels on the summary plot indicate 
gene sets that do not appear in the top 10 list of gene sets based on the "sort.by" 
argument (black labels) yet they appear in the top 5 list of gene sets based on 
the EGSEA "Significance Score". If two contrasts are provided, the rank is calculated 
based on the "comparison" analysis results and the "Significance Score" is calculated 
as the mean. If \code{sort.by = NULL}, the slot \code{sort.by} of the \code{object} 
is used to order gene sets.
The method \code{plotSummary} can be used to customize the Summary plots by 
changing the x-axis score
and filtering bubbles based on the values of the x-axis. The method \code{plotMethods} can be
used to generate Methods plots. \cr
4. Pathways page shows the KEGG pathways for the gene sets that are presented in the
Stats Table of a KEGG gene set collection. The gene fold changes are overlaid on the 
pathway maps and coloured based on the gene regulation direction: blue for down-regulation
and red for up-regulation. The method \code{plotPathway} can be used to generate
additional pathway maps. Note that this page only appears if a KEGG gene set collection
is used in the EGSEA analysis. \cr
5. Go Graphs page shows the Gene Ontology graphs for top 5 GO terms in each of 
three GO categories: Biological Processes (BP), Molecular Functions (MF), 
and Cellular Components (CC). Nodes are coloured based on the default \code{sort.by}
score where red indicates high significance and yellow indicates low significance. 
The method \code{plotGOGraph} can be used to customize GO graphs by 
changing the default sorting score and the number of significance nodes that can be
visualized. It is recommended that a small number of nodes is selected. Note that
this page only appears if a Gene Ontology gene set collection is used, i.e., for
the c5 collection from MSigDB or the gsdbgo collection from GeneSetDB. \cr
\cr
Finally, the "Interpret Results" hyperlink in the EGSEA report allows the user to download
the fold changes and limma analysis results and thus improve the interpretation of the results.
\cr 
Note that the running time of this function significantly increseas when 
\code{report = TRUE}. For example, the analysis in the example section below
was conducted on the $203$ signaling and disease KEGG pathways using a MacBook Pro 
machine that had a 2.8 GHz Intel Core i7 CPU and 16 GB of RAM. The execution time 
varied between 23.1 seconds (single thread) to 7.9 seconds (16 threads) when the HTML 
report generation was disabled. The execution time took 145.5 seconds when the report 
generation was enabled using 16 threads.
}
\examples{
# Example of egsea.cnt
library(EGSEAdata)
data(il13.data.cnt)
cnt = il13.data.cnt$counts
group = il13.data.cnt$group
design = il13.data.cnt$design
contrasts = il13.data.cnt$contra
genes = il13.data.cnt$genes
gs.annots = buildIdx(entrezIDs=rownames(cnt), species="human", 
msigdb.gsets="none",
         kegg.updated=FALSE, kegg.exclude = c("Metabolism"))
# set report = TRUE to generate the EGSEA interactive report
gsa = egsea.cnt(counts=cnt, group=group, design=design, contrasts=contrasts, 
         gs.annots=gs.annots, 
         symbolsMap=genes, baseGSEAs=egsea.base()[-c(2,5,6,9,12)], 
display.top = 5,
          sort.by="avg.rank", 
egsea.dir="./il13-egsea-cnt-report", 
         num.threads = 2, report = FALSE)
topSets(gsa) 

}
\references{
Monther Alhamdoosh, Milica Ng, Nicholas J. Wilson, Julie M. Sheridan, Huy Huynh, 
Michael J. Wilson, Matthew E. Ritchie; Combining multiple tools outperforms 
individual methods in gene set enrichment analyses. Bioinformatics 2017; 
33 (3): 414-424. doi: 10.1093/bioinformatics/btw623
}
\seealso{
\code{\link{topSets}}, \code{\link{egsea.base}}, 
\code{\link{egsea.sort}}, 
\code{\link{buildIdx}}, \code{\link{buildMSigDBIdx}}, 
\code{\link{buildKEGGIdx}},
\code{\link{buildGeneSetDBIdx}}, and \code{\link{buildCustomIdx}}
}

