% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egsea.R
\name{egsea.combine}
\alias{egsea.combine}
\title{EGSEA P-value Combining Options}
\usage{
egsea.combine()
}
\value{
It returns a character vector of available methods for the 
combineMethod argument in egsea
}
\description{
It lists the p-value combining methods
}
\examples{
egsea.combine()
}

