% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egsea.R
\name{egsea.sort}
\alias{egsea.sort}
\title{EGSEA Result Sorting Options}
\usage{
egsea.sort()
}
\value{
It returns a character vector of the accepted values for the sort.by 
argument in egsea
}
\description{
It lists the accepted sorting methods for analysis results
}
\examples{
egsea.sort()
}

