% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGSEAclass.R
\docType{class}
\name{EGSEAResults}
\alias{EGSEAResults}
\alias{EGSEAResults-class}
\title{The EGSEAResults class}
\description{
This class contains the results of an EGSEA analysis
}
\section{Slots}{

\describe{
\item{\code{results}}{list, EGSEA analysis results}

\item{\code{contrasts}}{character, the analysis contrasts}

\item{\code{sampleSize}}{numeric, number of samples}

\item{\code{gs.annots}}{list, the gene set collection annotation index}

\item{\code{baseMethods}}{character, vector of base GSE methods}

\item{\code{combineMethod}}{character, the p-value combining method}

\item{\code{sort.by}}{character, the results ordering argument}

\item{\code{symbolsMap}}{data.frame, the mapping between Entrez IDs and Gene Symbols}

\item{\code{logFC}}{matrix, the logFC matrix of contrasts}

\item{\code{report}}{logical, whether the report was generated}

\item{\code{report.dir}}{character, the directory of the EGSEA HTML report}
}}

