% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egsea.R
\docType{data}
\name{gsa}
\alias{gsa}
\title{EGSEA analysis results on the human IL-13 dataset}
\format{An object of class EGSEAResults}
\source{
The dataset of this analysis is available in the EGSEAdata package.
}
\description{
EGSEA analysis was performed on the il13.data from the EGSEAdata package using
the KEGG pathways, c2 and c5 gene set collections.
}

