% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSCollectionIndexclass.R
\docType{class}
\name{GSCollectionIndex}
\alias{$,GSCollectionIndex-method}
\alias{GSCollectionIndex}
\alias{GSCollectionIndex-class}
\alias{getSetByID}
\alias{getSetByID,GSCollectionIndex-method}
\alias{getSetByName}
\alias{getSetByName,GSCollectionIndex-method}
\alias{show,GSCollectionIndex-method}
\alias{summary,GSCollectionIndex-method}
\title{The GSCollectionIndex class}
\usage{
\S4method{$}{GSCollectionIndex}(x, name)

\S4method{summary}{GSCollectionIndex}(object)

\S4method{show}{GSCollectionIndex}(object)

getSetByName(object, set.name)

getSetByID(object, id)
}
\arguments{
\item{x}{GSCollectionIndex, the indexed gene set collection generated
from \code{\link{buildIdx}}, \code{\link{buildMSigDBIdx}}, 
\code{\link{buildKEGGIdx}},
\code{\link{buildGeneSetDBIdx}}, and \code{\link{buildCustomIdx}}.}

\item{name}{character, the slot name}

\item{object}{GSCollectionIndex, the indexed gene set collection generated
from \code{\link{buildIdx}}, \code{\link{buildMSigDBIdx}}, 
\code{\link{buildKEGGIdx}},
\code{\link{buildGeneSetDBIdx}}, and \code{\link{buildCustomIdx}}.}

\item{set.name}{character, a vector of gene set names as they appear in \code{\link{topSets}}.}

\item{id}{character, a vector of gene set IDs as they appears in the 
\code{\link{plotSummary}}.}
}
\value{
the selected slot.

summary of the collection.

details of the collection.

a list of the details of a gene set

a list of the details of a gene set
}
\description{
This class contains an indexed gene set collection

Extract a slot from an object of class GSCollectionIndex

A brief summary of the gene set collection

Print the details of the gene set collection

Print the details of a given gene set name

Print the details of a given gene set ID
}
\section{Slots}{

\describe{
\item{\code{original}}{list, the original gene sets}

\item{\code{idx}}{list, the gene set indexes}

\item{\code{anno}}{data.frame, the annotations of the gene sets}

\item{\code{featureIDs}}{character, vector of the original Entrez IDs that are 
used in the indexing procedure}

\item{\code{species}}{character, the species name}

\item{\code{name}}{character, the name of the gene set collection}

\item{\code{label}}{character, a label to distnguish this collection 
from other collections}
}}
\examples{
library(EGSEAdata)
data(il13.data)
v = il13.data$voom
gs.annots = buildIdx(entrezIDs=rownames(v$E), species="human", 
			msigdb.gsets="none", 
         kegg.updated=FALSE, kegg.exclude = c("Metabolism"))

print(gs.annots[[1]]$name)
library(EGSEAdata)
data(il13.data)
v = il13.data$voom
gs.annots = buildIdx(entrezIDs=rownames(v$E), species="human", 
			msigdb.gsets="none", 
         kegg.updated=FALSE, kegg.exclude = c("Metabolism"))

summary(gs.annots[[1]])
library(EGSEAdata)
data(il13.data)
v = il13.data$voom
gs.annots = buildIdx(entrezIDs=rownames(v$E), species="human", 
			msigdb.gsets="none", 
         kegg.updated=FALSE, kegg.exclude = c("Metabolism"))

show(gs.annots[[1]])
library(EGSEAdata)
data(il13.data)
v = il13.data$voom
gs.annots = buildIdx(entrezIDs=rownames(v$E), species="human", 
			msigdb.gsets="none", 
         kegg.updated=FALSE, kegg.exclude = c("Metabolism"))

getSetByName(gs.annots[[1]], "Asthma")
library(EGSEAdata)
data(il13.data)
v = il13.data$voom
gs.annots = buildIdx(entrezIDs=rownames(v$E), species="human", 
			msigdb.gsets="none", 
         kegg.updated=FALSE, kegg.exclude = c("Metabolism"))

getSetByID(gs.annots[[1]], "hsa04060")
}

