% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egsea.R
\name{plotMDS}
\alias{plotMDS}
\alias{plotMDS,EGSEAResults-method}
\title{Plot a multi-dimensional scaling (MDS) plot for the gene set rankings}
\usage{
plotMDS(object, gs.label = 1, contrast = 1, file.name = "methods.mds",
  format = "pdf", verbose = TRUE)
}
\arguments{
\item{object}{EGSEAResults object, the analysis result object from  \code{\link{egsea}}, 
\code{\link{egsea.cnt}}
or  \code{\link{egsea.ora}}.}

\item{gs.label}{character or numeric, the index/name of the KEGG pathways collection.
See names(object@gs.annots)[grep("^kegg", names(object@gs.annots))] for valid values.}

\item{contrast}{character or numeric, the index/name of the contrast or 0/"comparison". 
See object@contrasts for valid values.}

\item{file.name}{character, the name of the heatmap file without an extension.}

\item{format}{character, takes "pdf" or "png".}

\item{verbose}{logical, whether to print out progress messages and warnings.}
}
\description{
Plot a multi-dimensional scaling (MDS) plot for the gene set rankings
of the base GSE methods
}
\examples{
library(EGSEAdata)
data(il13.data)
v = il13.data$voom
contrasts = il13.data$contra
gs.annots = buildIdxEZID(entrezIDs=rownames(v$E), species="human", 
msigdb.gsets="none", 
         kegg.updated=FALSE, kegg.exclude = c("Metabolism"))
gsa = egsea(voom.results=v, contrasts=contrasts,  gs.annots=gs.annots, 
         symbolsMap=v$genes, 
baseGSEAs=egsea.base()[-c(2,5,6,9,12)], display.top = 5,
          sort.by="avg.rank", num.threads = 2, report=FALSE)
plotMDS(gsa)
}

