% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egsea.R
\name{plotSummary}
\alias{plotSummary}
\alias{plotSummary,EGSEAResults-method}
\title{Generate a Summary plot for EGSEA analysis}
\usage{
plotSummary(object, gs.label = 1, contrast = 1, file.name = "summary",
  format = "pdf", x.axis = "p.adj", x.cutoff = NULL, verbose = TRUE)
}
\arguments{
\item{object}{EGSEAResults object, the analysis result object from  \code{\link{egsea}}, 
\code{\link{egsea.cnt}}
or  \code{\link{egsea.ora}}.}

\item{gs.label}{character or numeric, the index/name of the KEGG pathways collection.
See names(object@gs.annots)[grep("^kegg", names(object@gs.annots))] for valid values.}

\item{contrast}{character or numeric, the index/name of the contrast. To compare two
contrasts, pass their indexes/names.
See object@contrasts for valid values.}

\item{file.name}{character, the name of the heatmap file without an extension.}

\item{format}{character, takes "pdf" or "png".}

\item{x.axis}{character, the x-axis of the summary plot. All the 
values accepted by the 
\strong{sort.by} parameter can be used. Default x.axis="p.value".}

\item{x.cutoff}{numeric, cut-off threshold to filter the gene sets of 
the summary plots
based on the values of the \strong{x.axis}. Default 
x.cutoff=NULL.}

\item{verbose}{logical, whether to print out progress messages and warnings.}
}
\description{
Generate a Summary plot for EGSEA analysis
}
\examples{
library(EGSEAdata)
data(il13.data)
v = il13.data$voom
contrasts = il13.data$contra
gs.annots = buildIdxEZID(entrezIDs=rownames(v$E), species="human", 
msigdb.gsets="none", 
         kegg.updated=FALSE, kegg.exclude = c("Metabolism"))
gsa = egsea(voom.results=v, contrasts=contrasts,  gs.annots=gs.annots, 
         symbolsMap=v$genes, 
baseGSEAs=egsea.base()[-c(2,5,6,9)], display.top = 5,
          sort.by="avg.rank", num.threads = 2, report=FALSE)
plotSummary(gsa)
plotSummary(gsa, contrast=c(1,2), file.name = "summary.cmp")

}

