% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egsea.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,EGSEAResults-method}
\title{Summary of the EGSEAResults object}
\usage{
\S4method{summary}{EGSEAResults}(object)
}
\arguments{
\item{object}{EGSEAResults object, the analysis result object from  \code{\link{egsea}}, 
\code{\link{egsea.cnt}}
or  \code{\link{egsea.ora}}.}
}
\description{
Summary of the EGSEAResults object
}
\examples{
library(EGSEAdata)
data(il13.data)
v = il13.data$voom
contrasts = il13.data$contra
gs.annots = buildIdxEZID(entrezIDs=rownames(v$E), species="human", 
msigdb.gsets="none", 
         kegg.updated=FALSE, kegg.exclude = c("Metabolism"))
gsa = egsea(voom.results=v, contrasts=contrasts,  gs.annots=gs.annots, 
         symbolsMap=v$genes, 
baseGSEAs=egsea.base()[-c(2,5,6,9)], display.top = 5,
          sort.by="avg.rank", num.threads = 2, report=FALSE)
summary(gsa)
}

