% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egsea.R
\name{topSets}
\alias{topSets}
\alias{topSets,EGSEAResults-method}
\title{Table of Top Gene Sets from an EGSEA Analysis}
\usage{
topSets(object, contrast = 1, gs.label = 1, sort.by = NULL, number = 10,
  names.only = TRUE, verbose = TRUE)
}
\arguments{
\item{object}{EGSEAResults object, the analysis result object from  \code{\link{egsea}}, 
\code{\link{egsea.cnt}}
or  \code{\link{egsea.ora}}.}

\item{contrast}{contrast column number or column name specifying which 
contrast is of interest.
if contrast = 0 or "comparison" and the number of contrasts greater than 1, 
the comparative gene sets are 
retruned.}

\item{gs.label}{the number or label of the gene set collection of interest.}

\item{sort.by}{character, determines how to order the analysis results in 
the stats table. 
The accepted values depend on the function used to generate the EGSEA 
results.}

\item{number}{integer, maximum number of gene sets to list}

\item{names.only}{logical, whether to display the EGSEA statistics or not.}

\item{verbose}{logical, whether to print out progress messages and warnings.}
}
\value{
A dataframe of top gene sets with the calculated statistics for each if 
names.only = FALSE.
}
\description{
Extract a table of the top-ranked gene sets from an EGSEA 
analysis.
}
\examples{
library(EGSEAdata)
data(il13.data)
v = il13.data$voom
contrasts = il13.data$contra
gs.annots = buildIdxEZID(entrezIDs=rownames(v$E), species="human", 
msigdb.gsets="none", 
         kegg.updated=FALSE, kegg.exclude = c("Metabolism"))
gsa = egsea(voom.results=v, contrasts=contrasts,  gs.annots=gs.annots, 
         symbolsMap=v$genes, 
baseGSEAs=egsea.base()[-c(2,5,6,9)], display.top = 5,
          sort.by="avg.rank",  
         num.threads = 2, report=FALSE)
topSets(gsa, contrast=1, gs.label="kegg", number = 10)
topSets(gsa, contrast=1, gs.label=1, sort.by="ora", number = 10, 
names.only=FALSE)
topSets(gsa, contrast=0, gs.label="kegg", number = 10)

}

