% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egseaMain.R
\docType{package}
\name{EGSEA-package}
\alias{EGSEA-package}
\alias{EGSEA}
\title{Ensemble of Gene Enrichment Analysis (EGSEA)}
\description{
This packages provides the implementation of the EGSEA 
algorithm and addition functions to help perform GSE analysis.
}
\references{
Monther Alhamdoosh, Milica Ng, Nicholas J. Wilson, Julie M. Sheridan, Huy Huynh, 
Michael J. Wilson, Matthew E. Ritchie; Combining multiple tools outperforms 
individual methods in gene set enrichment analyses. Bioinformatics 2017; 
33 (3): 414-424. doi: 10.1093/bioinformatics/btw623
}
\author{
Monther Alhamdoosh, Milica Ng and Matthew Ritchie
}
