% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egsea.R
\docType{package}
\name{EGSEA-package}
\alias{EGSEA}
\alias{EGSEA-package}
\alias{egsea.logo}
\title{Ensemble of Gene Enrichment Analysis (EGSEA)}
\usage{
egsea.logo(out.dir = "./")
}
\arguments{
\item{out.dir}{character, the target directory to which the logo will be written.}
}
\value{
a PNG file.
}
\description{
This packages provides the implementation of the EGSEA 
algorithm and addition functions to help perform GSE analysis.

This function writes out the official EGSEA package logo
}
\details{
This function generates a PNG file of the EGSEA logo, which can
be used to acknowledge EGSEA in presentations/reports. 
The logo was designed by Roberto Bonelli from The Walter and Eliza Hall Institute
of Medical Research.
}
\author{
Monther Alhamdoosh, Milica Ng and Matthew Ritchie
}
\references{
Monther Alhamdoosh, Milica Ng, Nicholas J. Wilson, Julie M. Sheridan, Huy Huynh, 
Michael J. Wilson, Matthew E. Ritchie; Combining multiple tools outperforms 
individual methods in gene set enrichment analyses. Bioinformatics 2017; 
33 (3): 414-424. doi: 10.1093/bioinformatics/btw623
}

