% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGSEAclass.R
\docType{class}
\name{EGSEAResults}
\alias{$,EGSEAResults-method}
\alias{EGSEAResults}
\alias{EGSEAResults-class}
\alias{plotGOGraph}
\alias{plotGOGraph,EGSEAResults-method}
\alias{plotHeatmap}
\alias{plotHeatmap,EGSEAResults-method}
\alias{plotMethods}
\alias{plotMethods,EGSEAResults-method}
\alias{plotPathway}
\alias{plotPathway,EGSEAResults-method}
\alias{plotSummary}
\alias{plotSummary,EGSEAResults-method}
\alias{show,EGSEAResults-method}
\alias{showSetByID}
\alias{showSetByID,EGSEAResults-method}
\alias{showSetByName}
\alias{showSetByName,EGSEAResults-method}
\alias{summary,EGSEAResults-method}
\alias{topSets}
\alias{topSets,EGSEAResults-method}
\title{The EGSEAResults class}
\usage{
\S4method{$}{EGSEAResults}(x, name)

topSets(object, gs.label = 1, contrast = 1, sort.by = NULL, number = 10,
  names.only = TRUE, verbose = TRUE)

\S4method{show}{EGSEAResults}(object)

\S4method{summary}{EGSEAResults}(object)

plotHeatmap(object, gene.set, gs.label = 1, contrast = 1,
  file.name = "heatmap", format = "pdf", verbose = TRUE)

plotPathway(object, gene.set, gs.label = 1, contrast = 1,
  file.name = "pathway", verbose = TRUE)

plotMethods(object, gs.label = 1, contrast = 1, file.name = "methods.mds",
  format = "pdf", verbose = TRUE)

plotSummary(object, gs.label = 1, contrast = 1, file.name = "summary",
  format = "pdf", x.axis = "p.adj", x.cutoff = NULL, verbose = TRUE)

plotGOGraph(object, gs.label = "c5", contrast = 1, sort.by = "p.value",
  noSig = 5, file.name = "c5-top-", format = "pdf", verbose = TRUE)

showSetByName(object, gs.label = 1, set.name)

showSetByID(object, gs.label = 1, id)
}
\arguments{
\item{x}{EGSEAResults object, the analysis result object from  \code{\link{egsea}}, 
\code{\link{egsea.cnt}}
or  \code{\link{egsea.ora}}.}

\item{name}{character, the slot name}

\item{object}{EGSEAResults object, the analysis result object from  \code{\link{egsea}}, 
\code{\link{egsea.cnt}}
or  \code{\link{egsea.ora}}.}

\item{gs.label}{the number or label of the gene set collection of interest.}

\item{contrast}{contrast column number or column name specifying which 
contrast is of interest.
if contrast = 0 or "comparison" and the number of contrasts greater than 1, 
the comparative gene sets are 
retruned.}

\item{sort.by}{character, determines how to order the analysis results in 
the stats table. 
The accepted values depend on the function used to generate the EGSEA 
results.}

\item{number}{integer, maximum number of gene sets to list}

\item{names.only}{logical, whether to display the EGSEA statistics or not.}

\item{verbose}{logical, whether to print out progress messages and warnings.}

\item{gene.set}{character, the name of the gene set. 
See the output of \code{\link{topSets}}.}

\item{file.name}{character, the prefix of the output file name.}

\item{format}{character, takes "pdf" or "png".}

\item{x.axis}{character, the x-axis of the summary plot. All the 
values accepted by the 
\strong{sort.by} parameter can be used. Default x.axis="p.value".}

\item{x.cutoff}{numeric, cut-off threshold to filter the gene sets of 
the summary plots
based on the values of the \strong{x.axis}. Default 
x.cutoff=NULL.}

\item{noSig}{numeric, number of significant GO terms to be displayed. A number larger than 
5 might not work due to the size of the generated graph.}

\item{set.name}{character, a vector of gene set names as they appear in \code{\link{topSets}}.}

\item{id}{character, a vector of gene set IDs as they appears in the 
\code{\link{plotSummary}}.}
}
\value{
the selected slot

A dataframe of top gene sets with the calculated statistics for each if 
names.only = FALSE.

nothing.

nothing.

Heatmap plot.

Pathway map plot.

MDS plot

Summary plot

GO graphs of sginificant BP, MF and CC terms.

details of a gene set

details of a gene set
}
\description{
This class contains the results of an EGSEA analysis

Extract a slot from an object of class EGSEAResults

Extract a table of the top-ranked gene sets from an EGSEA 
analysis.

Show the parameters of the EGSEAResults object

Display a brief summary of the analysis of the EGSEAResults object

Generate a heatmap of fold changes for a selected gene set

Generate a visual map for a selected KEGG pathway

Generate a multi-dimensional scaling (MDS) plot for the gene set rankings
of the base GSE methods

Generate a Summary plot for EGSEA analysis

Generate a graph of the top significant GO terms in a GO term collection.
It could be c5 from MSigDB or Gene Ontolog from the GeneSetDB.

Print the details of a given gene set name

Print the details of a given gene set ID
}
\section{Slots}{

\describe{
\item{\code{results}}{list, EGSEA analysis results}

\item{\code{contrasts}}{character, the analysis contrasts}

\item{\code{sampleSize}}{numeric, number of samples}

\item{\code{gs.annots}}{list, the gene set collection annotation index}

\item{\code{baseMethods}}{character, vector of base GSE methods}

\item{\code{combineMethod}}{character, the p-value combining method}

\item{\code{sort.by}}{character, the results ordering argument}

\item{\code{symbolsMap}}{data.frame, the mapping between Entrez IDs and Gene Symbols}

\item{\code{logFC}}{matrix, the logFC matrix of contrasts}

\item{\code{report}}{logical, whether the report was generated}

\item{\code{report.dir}}{character, the directory of the EGSEA HTML report}
}}
\examples{
data(il13.gsa)
class(gsa)
print(gsa$baseMethods)
data(il13.gsa)
class(gsa)
topSets(gsa,  gs.label="kegg",contrast=1, number = 10)
topSets(gsa,  gs.label=1, contrast=1, sort.by="ora", number = 10, 
names.only=FALSE)
topSets(gsa,  gs.label="kegg",contrast=0, number = 10)

data(il13.gsa)
class(gsa)
show(gsa)
data(il13.gsa)
class(gsa)
summary(gsa)
data(il13.gsa)
class(gsa)
plotHeatmap(gsa, gs.label="kegg", "Asthma")
plotHeatmap(gsa, gs.label="kegg", "Asthma", contrast = "comparison", 
file.name = "asthma.hm.cmp")
data(il13.gsa)
class(gsa)
plotPathway(gsa, gs.label="kegg", "Asthma")
plotPathway(gsa, gs.label="kegg", "Asthma", contrast="comparison", 
file.name = "asthma.map.cmp")
data(il13.gsa)
class(gsa)
plotMethods(gsa)
data(il13.gsa)
class(gsa)
plotSummary(gsa)
plotSummary(gsa, contrast=c(1,2), file.name = "summary.cmp")

data(il13.gsa)
class(gsa)
plotGOGraph(gsa, sort.by="avg.rank")
data(il13.gsa)
class(gsa)
showSetByName(gsa, "kegg", "Asthma")
data(il13.gsa)
class(gsa)
showSetByID(gsa, "kegg", "hsa04060")
}

